[core] Move coordinator impl into modules
